virtio-mmio: Refactor common functionality
Common functionality is refactored into virtio_mmio_common.h
in order to MSI support in later patch set.
